Shifted voting control into a PureComponent child to avoid repainting bugs
